sui_indexer_alt_graphql
Function start_rpc

pub async fn start_rpc(
    database_url: Option<Url>,
    bigtable_instance: Option<String>,
    db_args: DbArgs,
    bigtable_args: BigtableArgs,
    args: RpcArgs,
    system_package_task_args: SystemPackageTaskArgs,
    version: &'static str,
    config: RpcConfig,
    pg_pipelines: Vec<String>,
    registry: &Registry,
    cancel: CancellationToken,
) -> Result<JoinHandle<()>>

Set-up and run the RPC service, using the provided arguments (expected to be extracted from the command-line). The service will continue to run until the cancellation token is triggered, and will signal cancellation on the token when it is shutting down.

Access to most reads is controlled by the database_url – if it is None, those reads will not work. KV queries can optionally be served by a Bigtable instance, if bigtable_instance is provided, otherwise these requests are served by the database. If a bigtable_instance is provided, the GOOGLE_APPLICATION_CREDENTIALS environment variable must point to the credentials JSON file.

version is the version string reported in response headers by the service as part of every request.

The service may spin up auxiliary services (such as the system package task) to support itself, and will clean these up on shutdown as well.